refactor: atoms/buttonのディレクトリ構成を改善
