+++ src/atoms/button/charAnimationButton/index.module.scss
