+++ src/atoms/button/iconButton/index.tsx
@@ -1,4 +1,4 @@
-import styles from "./iconButton.module.scss";
+import styles from "./index.module.scss";
 
 interface Props {
   imgLink: string;
